src/workers/export-worker/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
minio>=7.2.0
python-docx>=1.1.2
